refactor: organize services
