+++ app/routes/_app.reimbursements.$reimbursementId.tsx
@@ -24,7 +24,7 @@ import { TransactionItemMethod, TransactionItemType } from "~/lib/constants";
 import { getPrismaErrorText, notFound } from "~/lib/responses.server";
 import { toast } from "~/lib/toast.server";
 import { capitalize, formatCentsAsDollars } from "~/lib/utils";
-import { MailService } from "~/services.server/MailService.server";
+import { MailService } from "~/services.server/mail";
 import { SessionService } from "~/services.server/session";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => [
