+++ app/services.server/auth.ts
@@ -1,12 +1,26 @@
-import type { Password, User } from "@prisma/client";
+import { Password, User } from "@prisma/client";
 import bcrypt from "bcryptjs";
 
 import { db } from "~/integrations/prisma.server";
 
-export async function verifyLogin(username: NonNullable<User["username"]>, password: Password["hash"]) {
+export function hashPassword(password: string) {
+  return bcrypt.hash(password, 10);
+}
+
+export function comparePasswords(password: string, hash: string) {
+  return bcrypt.compare(password, hash);
+}
+
+export async function verifyLogin({
+  username,
+  password,
+}: {
+  username: NonNullable<User["username"]>;
+  password: Password["hash"];
+}) {
   const userWithPassword = await db.user.findUnique({
     where: {
-      // not deactivated
+      // Not deactivated
       isActive: true,
       username,
     },
@@ -20,7 +34,7 @@ export async function verifyLogin(username: NonNullable<User["username"]>, passw
     return null;
   }
 
-  const isValid = await bcrypt.compare(password, userWithPassword.password.hash);
+  const isValid = await comparePasswords(password, userWithPassword.password.hash);
 
   if (!isValid) {
     return null;
